+++ src/App.js
@@ -1,21 +1,7 @@
 import React from 'react';
+import TodoList from './TodoList'
+import AddTodoForm from './AddTodoForm'
 import './style.css'
-
-
-const todoList = [
-  {
-    'id': 1,
-    'title': 'Complete the coding challenge on uidaily.com'
-  },
-  {
-    'id': 2,
-    'title': 'Search for a remote job'
-  },
-  {
-    'id': 3,
-    'title': 'Ask for help moving furniture next month'
-  },
-]
 
 function App() {
   return (
@@ -23,19 +9,10 @@ function App() {
       <header>
         <h1>Todo List</h1>
       </header>
-        <List />
+        <AddTodoForm />
+        <TodoList />
     </div>
   );
 }
-
-const List = () => (
-  <ul>
-  {todoList.map((item) => (
-    <li key={item.id}>
-      <span>{item.title}</span>
-    </li>
-  ))}
-  </ul>
-)
 
 export default App;
